requirements.txt
streamlit
requests
pandas
matplotlib
seaborn
python-docx
